grafana/data: Move useFieldOverrides from core
