+++ packages/grafana-data/package.json
@@ -49,6 +49,7 @@
     "moment-timezone": "0.5.38",
     "ol": "7.1.0",
     "papaparse": "5.3.2",
+    "react-use": "17.4.0",
     "regenerator-runtime": "0.13.10",
     "rxjs": "7.5.7",
     "tinycolor2": "1.4.2",
